Apply Styleguide Rules 04-* Application Structure (#858)
* Apply Styleguide Rules 04-* Application Structure

Applies the Styleguide rules for the application structure.
Fixes #857.

* Update barrel imports

Updates the barrel imports to the `/index` syntax. Furthermore,
corrects the template and style urls of the components.

* Update README

Updates the directory structure in the README to reflect the
recent changes to the application folder structure to conform
to the styleguide reccomendations.