Clean up (#1041)
* Add Environment Configuration to TOC

Adds a link to the environment configuration to the table of contents.

* Reorder angular packages alphabetically

Reorders the angular packages alphabetically in the seed.config.ts
(SystemJS, System Builder) and test-main.js to be in accordance to
order in the package.json.

* Update Route Configuration for lazy loaded components

Updates the route configuration for the +about and +home lazy loaded
component according to the angular guide
(see: https://angular.io/docs/ts/latest/guide/router.html).
Furthermore removes the index property which was deprecated with
@angular/router 3.0.0-alpha.7. Now, the `path: ''` indicates the index
of a route.

* Remove workaround for angular upstream issue in index.html

Removes the workaround present in the index.html that was made because
of the angular issue #6501
(https://github.com/angular/angular/issues/6501).
For information on the fix, see here:
https://github.com/angular/angular/issues/6501#issuecomment-220735527